requirements.txt
fastapi
fastapi[standard]
uvicorn
psycopg2-binary